Check the certificate for the correct domain.
If a server offers XMPP-over-TLS on port 443 it is often on a different machine, e.g. the user@example.org connects to xmpp.example.org:443. The certificate must be valid for example.org not for xmpp.example.org but the lib always checked for the domain of the host we connect to (xmpp.example.org).